Reorganize crag components file structure and naming.
