+++ src/components/crag/crag-comments.tsx
@@ -1,0 +1,5 @@
+function CragComments() {
+  return <div>Hello crag comments!</div>;
+}
+
+export default CragComments;
