+++ src/components/crag/crag-gallery.tsx
@@ -1,0 +1,5 @@
+function CragGallery() {
+  return <div>Hello crag gallery!</div>;
+}
+
+export default CragGallery;
