+++ src/components/crag/crag-info.tsx
@@ -1,0 +1,5 @@
+function CragInfo() {
+  return <div>Hello crag info!</div>;
+}
+
+export default CragInfo;
