+++ src/components/crag/crag-routes/crag-route-list/crag-route.tsx
@@ -1,15 +1,15 @@
 import { useContext } from "react";
-import { Crag, Route } from "../../graphql/generated";
-import Grade from "../grade";
-import RouteLink from "../route-link";
-import AscentIcon from "../ui/ascent-icon";
-import Checkbox from "../ui/checkbox";
-import IconComment from "../ui/icons/comment";
-import { IconSize } from "../ui/icons/icon";
-import IconStarEmpty from "../ui/icons/star-empty";
-import IconStarFull from "../ui/icons/star-full";
-import Link from "../ui/link";
-import { CragTableContext } from "./crag-table";
+import { Crag, Route } from "../../../../graphql/generated";
+import Grade from "../../../grade";
+import RouteLink from "../../../route-link";
+import AscentIcon from "../../../ui/ascent-icon";
+import Checkbox from "../../../ui/checkbox";
+import IconComment from "../../../ui/icons/comment";
+import { IconSize } from "../../../ui/icons/icon";
+import IconStarEmpty from "../../../ui/icons/star-empty";
+import IconStarFull from "../../../ui/icons/star-full";
+import Link from "../../../ui/link";
+import { CragTableContext } from "../../crag-routes";
 
 interface Props {
   crag: Crag;
